# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Trivela API
  version: 0.1.0
  description: |
    REST API for the Trivela campaign and rewards platform. Handles campaign metadata,
    health checks, and Soroban RPC configuration for the frontend.

    See the [API Migration Guide](../docs/API_MIGRATION.md) for v0 → v1 breaking changes,
    endpoint changelog, deprecation policy, and migration examples with curl snippets.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:3001
    description: Local development server
  - url: https://api.trivela.example.com
    description: Production server
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /api/v1/campaigns/{id}:
    get:
      summary: Get campaign by ID
      tags:
        - Campaigns
      operationId: getCampaignById
      security: []
      parameters:
        - name: id
          in: path
          required: true
          description: Campaign ID
          schema:
            type: string
      responses:
        '200':
          description: Campaign details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '404':
          description: Campaign not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update campaign
      tags:
        - Campaigns
      operationId: updateCampaign
      parameters:
        - name: id
          in: path
          required: true
          description: Campaign ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampaignUpdate'
      responses:
        '200':
          description: Campaign updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Campaign not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete campaign
      tags:
        - Campaigns
      operationId: deleteCampaign
      parameters:
        - name: id
          in: path
          required: true
          description: Campaign ID
          schema:
            type: string
      responses:
        '204':
          description: Campaign deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Campaign not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/campaigns/{id}/stats:
    get:
      summary: Get campaign analytics and stats
      tags:
        - Campaigns
      operationId: getCampaignStats
      security: []
      parameters:
        - name: id
          in: path
          required: true
          description: Campaign ID
          schema:
            type: string
        - name: range
          in: query
          description: Time range (7d, 30d, all)
          schema:
            type: string
            enum: [7d, 30d, all]
            default: 7d
        - name: from
          in: query
          description: Start date (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: End date (ISO 8601)
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Campaign statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignStats'
        '404':
          description: Campaign not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/index/campaigns/{id}/stats:
    get:
      summary: Rollup stats for a campaign (ETag-cached)
      tags: [IndexRead]
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Campaign stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  campaignId: { type: string }
                  as_of_ledger: { type: integer, nullable: true }
                  summary:
                    type: object
                    properties:
                      totalParticipants: { type: integer }
                      totalCredited: { type: string }
                      totalClaimed: { type: string }
                      claimRate: { type: number }
        '304':
          description: Not Modified (ETag match)
        '404':
          description: Campaign not found

  # ── #556: Sponsored accounts ───────────────────────────────────────────────
...
  /api/v1/campaigns/{id}/claimable-balances:
    post:
      summary: Trigger claimable balance creation for eligible-unclaimed users
      tags: [ClaimableBalances]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                graceDays:
                  type: integer
                  default: 30
                assetCode:
                  type: string
                assetIssuer:
                  type: string
      responses:
        '202':
          description: >-
            Submission enqueued on the durable job queue; returns { ok, campaignId, jobId, status:
            "queued" } immediately. Poll GET on this same path to see rows move pending ->
            created/failed.
        '404':
          description: Campaign not found
    get:
      summary: List claimable balances for a campaign
      tags: [ClaimableBalances]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: status
          schema:
            {
              type: string,
              enum: [pending, created, claimed_by_user, reclaimed_by_operator, failed],
            }
        - in: query
          name: limit
          schema: { type: integer, default: 50, maximum: 200 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: Paginated claimable balance list
  /api/v1/campaigns/{id}/claimable-balances/reclaim:
    post:
      summary: Operator reclaim of expired unclaimed balances
      tags: [ClaimableBalances]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Balances marked as reclaimed
        '404':
          description: Campaign not found
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for write operations
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication
  schemas:
    Campaign:
      type: object
      required:
        - id
        - name
        - slug
        - description
        - active
        - featured
        - rewardPerAction
        - createdAt
        - updatedAt
        - status
      properties:
        id:
          type: string
          description: Campaign ID
        name:
          type: string
          description: Campaign name
        slug:
          type: string
          description: URL-friendly slug
        description:
          type: string
          description: Campaign description
        active:
          type: boolean
          description: Whether campaign is active
        featured:
          type: boolean
          description: Whether campaign is featured
        hidden:
          type: boolean
          description: Whether campaign is hidden
        hiddenReason:
          type: string
          nullable: true
          description: Reason for hiding campaign
        rewardPerAction:
          type: number
          description: Reward points per action
        startDate:
          type: string
          format: date-time
          nullable: true
          description: Campaign start date
        endDate:
          type: string
          format: date-time
          nullable: true
          description: Campaign end date
        status:
          type: string
          enum: [draft, published, archived]
          description: Campaign publish-workflow status
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
        imageUrl:
          type: string
          format: uri
          nullable: true
          description: Public URL for campaign image
        tags:
          type: array
          items:
            type: string
          description: Campaign tags
        category:
          type: string
          nullable: true
          description: Campaign category
...
    CampaignUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          description: Campaign name
        slug:
          type: string
          description: URL-friendly slug
        description:
          type: string
          description: Campaign description
        rewardPerAction:
          type: number
          minimum: 0
          description: Reward points per action
        active:
          type: boolean
          description: Whether campaign is active
        featured:
          type: boolean
          description: Whether campaign is featured
        hidden:
          type: boolean
          description: Whether campaign is hidden
        hiddenReason:
          type: string
          nullable: true
          description: Reason for hiding campaign
        startDate:
          type: string
          format: date-time
          nullable: true
          description: Campaign start date
        endDate:
          type: string
          format: date-time
          nullable: true
          description: Campaign end date
...
    Error:
      type: object
      required:
        - error
        - code
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
    ValidationError:
      type: object
      required:
        - error
        - code
        - details
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: array
          items:
            type: string
          description: Validation error details
...
    CampaignStats:
      type: object
      required:
        - campaignId
        - onChainSynced
        - range
        - summary
        - registrationsByDay
        - pointsByDay
      properties:
        campaignId:
          type: string
          description: Campaign ID
        onChainSynced:
          type: boolean
          description: Whether on-chain data has been synced
        range:
          type: object
          properties:
            from:
              type: string
              format: date-time
            to:
              type: string
              format: date-time
        summary:
          type: object
          required:
            - totalParticipants
            - totalPoints
            - claimRate
            - timeRemainingMs
          properties:
            totalParticipants:
              type: integer
              description: Total registered participants
            totalPoints:
              type: integer
              description: Total points awarded
            claimRate:
              type: number
              description: Percentage of points claimed
            timeRemainingMs:
              type: integer
              description: Milliseconds until campaign ends
        registrationsByDay:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              count:
                type: integer
          description: Daily registration counts
        pointsByDay:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              credited:
                type: integer
              claimed:
                type: integer
          description: Daily credited and claimed points